Refactoring. Split validators to separated classes. Introduced strategy class for file validation. Small changes.
