+++ utils/excel_utils.py
@@ -1,10 +1,11 @@
 import pandas as pd
+from pandas import DataFrame
 
 __EXCEL_FILE_EXTENSION = 'xlsx'
 
 
 # Read Excel file as DataFrame
-def get_file_as_data_frame(path: str, sheet_name: str = None) -> pd.DataFrame:
+def get_file_as_data_frame(path: str, sheet_name: str = None) -> DataFrame:
     if sheet_name is None:
         return pd.read_excel(path, engine='openpyxl')
     else:
@@ -12,5 +13,5 @@ def get_file_as_data_frame(path: str, sheet_name: str = None) -> pd.DataFrame:
 
 
 # Save a dataframe to selected path as .xlsx file
-def save_data_data_frame_as_excel_file_to_path(data_frame: pd.DataFrame, path_to_save: str) -> None:
+def save_data_data_frame_as_excel_file_to_path(data_frame: DataFrame, path_to_save: str) -> None:
     data_frame.to_excel(f'{path_to_save}.{__EXCEL_FILE_EXTENSION}', index=False)
